# Implement PreScore and change NodeNumberPlugin to use it