--- a/guest/testdata/all/main.go
+++ b/guest/testdata/all/main.go
@@ -31,7 +31,7 @@ import (
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/score"
 )
 
-type all interface {
+type extensionPoints interface {
 	api.EnqueueExtensions
 	api.PreFilterPlugin
 	api.FilterPlugin
@@ -41,11 +41,11 @@ type all interface {
 func main() {
 	// Multiple tests are here to reduce re-compilation time and size checked
 	// into git.
-	var plugin all
+	var plugin extensionPoints
 	if len(os.Args) == 2 && os.Args[1] == "params" {
-		plugin = params{}
+		plugin = paramsPlugin{}
 	} else {
-		plugin = noop{}
+		plugin = noopPlugin{}
 	}
 
 	enqueue.SetPlugin(plugin)
@@ -54,44 +54,53 @@ func main() {
 	score.SetPlugin(plugin)
 }
 
-// noop doesn't do anything, and this style isn't recommended. This shows the
+// noopPlugin doesn't do anything, and this style isn't recommended. This shows the
 // impact two things:
 //
 //   - implementing multiple interfaces
 //   - overhead of constructing function parameters
-type noop struct{}
+type noopPlugin struct{}
 
-func (noop) EventsToRegister() (clusterEvents []api.ClusterEvent) { return }
+func (noopPlugin) EventsToRegister() (clusterEvents []api.ClusterEvent) { return }
 
-func (noop) PreFilter(api.CycleState, proto.Pod) (nodeNames []string, status *api.Status) { return }
-
-func (noop) Filter(api.CycleState, proto.Pod, api.NodeInfo) (status *api.Status) { return }
-
-func (noop) Score(api.CycleState, proto.Pod, string) (score int32, status *api.Status) { return }
-
-// params doesn't do anything, except evaluate each parameter. This shows if
-// protobuf unmarshal caching works (for the pod), and also baseline
-// performance of reading each parameter.
-type params struct{}
-
-func (params) EventsToRegister() (clusterEvents []api.ClusterEvent) {
+func (noopPlugin) PreFilter(api.CycleState, proto.Pod) (nodeNames []string, status *api.Status) {
 	return
 }
 
-func (params) PreFilter(state api.CycleState, pod proto.Pod) (nodeNames []string, status *api.Status) {
+func (noopPlugin) Filter(api.CycleState, proto.Pod, api.NodeInfo) (status *api.Status) { return }
+
+func (noopPlugin) Score(api.CycleState, proto.Pod, string) (score int32, status *api.Status) { return }
+
+// paramsPlugin doesn't do anything, except evaluate each parameter. This shows
+// if protobuf unmarshal caching works (for the pod), and also baseline
+// performance of reading each parameter.
+type paramsPlugin struct{}
+
+func (paramsPlugin) EventsToRegister() (clusterEvents []api.ClusterEvent) {
+	return
+}
+
+func (paramsPlugin) PreFilter(state api.CycleState, pod proto.Pod) (nodeNames []string, status *api.Status) {
 	_, _ = state.Read("ok")
 	_ = pod.Spec()
 	return
 }
 
-func (params) Filter(state api.CycleState, pod proto.Pod, nodeInfo api.NodeInfo) (status *api.Status) {
+func (paramsPlugin) Filter(state api.CycleState, pod proto.Pod, nodeInfo api.NodeInfo) (status *api.Status) {
 	_, _ = state.Read("ok")
 	_ = pod.Spec()
 	_ = nodeInfo.Node().Spec() // trigger lazy loading
 	return
 }
 
-func (params) Score(state api.CycleState, pod proto.Pod, nodeName string) (score int32, status *api.Status) {
+func (paramsPlugin) PreScore(state api.CycleState, pod proto.Pod, nodeList proto.NodeList) *api.Status {
+	_, _ = state.Read("ok")
+	_ = pod.Spec()
+	_ = nodeList.Items()
+	return nil
+}
+
+func (paramsPlugin) Score(state api.CycleState, pod proto.Pod, nodeName string) (score int32, status *api.Status) {
 	_, _ = state.Read("ok")
 	_ = pod.Spec()
 	_ = nodeName