--- a/guest/testdata/cyclestate/main.go
+++ b/guest/testdata/cyclestate/main.go
@@ -14,6 +14,8 @@
    limitations under the License.
 */
 
+// Package main tests state propagation through defined extension points.
+// See https://kubernetes.io/docs/concepts/scheduling-eviction/scheduling-framework/#extension-points
 package main
 
 // Override the default GC with a more performant one.
@@ -29,6 +31,7 @@ import (
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/enqueue"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/filter"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/prefilter"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/prescore"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/score"
 	protoapi "sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto/api"
 )
@@ -53,65 +56,106 @@ func main() {
 		}
 	}
 
-	plugin := cyclestate{}
+	plugin := statePlugin{}
 	enqueue.SetPlugin(plugin)
 	prefilter.SetPlugin(plugin)
 	filter.SetPlugin(plugin)
+	prescore.SetPlugin(plugin)
 	score.SetPlugin(plugin)
 }
 
 const (
 	// name is the name of the plugin used in the plugin registry and configurations.
 	name              = "CycleState"
-	prefilterStateKey = "PreFilter" + name
+	preFilterStateKey = "PreFilter" + name
+	preScoreStateKey  = "PreScore" + name
 )
 
-// cyclestate makes sure api.CycleState is consistent between callbacks.
-type cyclestate struct{}
+// statePlugin makes sure api.CycleState is consistent between callbacks.
+type statePlugin struct{}
 
 var clusterEvents []api.ClusterEvent
 
-func (cyclestate) EventsToRegister() []api.ClusterEvent { return clusterEvents }
+func (statePlugin) EventsToRegister() []api.ClusterEvent { return clusterEvents }
 
 // podSpec is used to test cycle state coherency.
 var podSpec *protoapi.PodSpec
 
-type prefilterStateVal map[string]any
+type preFilterStateVal map[string]any
 
-func (cyclestate) PreFilter(state api.CycleState, pod proto.Pod) (nodeNames []string, status *api.Status) {
+type preScoreStateVal map[string]any
+
+func (statePlugin) PreFilter(state api.CycleState, pod proto.Pod) (nodeNames []string, status *api.Status) {
 	if nextPodSpec := pod.Spec(); unsafe.Pointer(nextPodSpec) == unsafe.Pointer(podSpec) {
 		panic("didn't reset pod on pre-filter")
 	} else {
 		podSpec = nextPodSpec
 	}
-	if _, ok := state.Read(prefilterStateKey); ok {
-		panic("didn't reset state on pre-filter")
+	mustNotScoreState(state)
+	if _, ok := state.Read(preFilterStateKey); ok {
+		panic("didn't reset filter state on pre-filter")
 	} else {
-		state.Write(prefilterStateKey, prefilterStateVal{})
+		state.Write(preFilterStateKey, preFilterStateVal{})
 	}
 	return
 }
 
-func (cyclestate) Filter(state api.CycleState, pod proto.Pod, _ api.NodeInfo) (status *api.Status) {
+func (statePlugin) Filter(state api.CycleState, pod proto.Pod, _ api.NodeInfo) (status *api.Status) {
 	if unsafe.Pointer(pod.Spec()) != unsafe.Pointer(podSpec) {
 		panic("didn't cache pod from pre-filter")
 	}
-	if val, ok := state.Read(prefilterStateKey); !ok {
+	mustNotScoreState(state)
+	if val, ok := state.Read(preFilterStateKey); !ok {
+		panic("didn't propagate filter state from pre-filter")
+	} else {
+		val.(preFilterStateVal)["filter"] = struct{}{}
+	}
+	return
+}
+
+func (statePlugin) PreScore(state api.CycleState, pod proto.Pod, _ proto.NodeList) *api.Status {
+	if unsafe.Pointer(pod.Spec()) != unsafe.Pointer(podSpec) {
+		panic("didn't cache pod from filter")
+	}
+	mustFilterState(state)
+	if _, ok := state.Read(preScoreStateKey); ok {
+		panic("didn't reset score state on pre-score")
+	} else {
+		state.Write(preScoreStateKey, preScoreStateVal{})
+	}
+	return nil
+}
+
+func (statePlugin) Score(state api.CycleState, pod proto.Pod, _ string) (score int32, status *api.Status) {
+	if unsafe.Pointer(pod.Spec()) != unsafe.Pointer(podSpec) {
+		panic("didn't cache pod from pre-score")
+	}
+	mustFilterState(state)
+	if val, ok := state.Read(preScoreStateKey); !ok {
+		panic("didn't propagate score state from pre-score")
+	} else {
+		val.(preScoreStateVal)["score"] = struct{}{}
+	}
+	return
+}
+
+// mustNotScoreState ensures that score state, written after filter, cannot
+// be read by extension points before it.
+//
+// Note: Tests will need to be revisited when plugins become re-entrant for
+// reasons such as preemption!
+func mustNotScoreState(state api.CycleState) {
+	if _, ok := state.Read(preScoreStateKey); ok {
+		panic("didn't reset score state on pre-filter")
+	}
+}
+
+// mustFilterState ensures that score, which happens after filter, can still
+// see state written before it.
+func mustFilterState(state api.CycleState) {
+	if val, ok := state.Read(preFilterStateKey); !ok {
 		panic("didn't propagate state from pre-filter")
-	} else {
-		val.(prefilterStateVal)["a"] = struct{}{}
+	} else if _, ok = val.(preFilterStateVal)["filter"]; !ok {
+		panic("filter value lost propagating from pre-score")
 	}
-	return
-}
-
-func (cyclestate) Score(state api.CycleState, pod proto.Pod, _ string) (score int32, status *api.Status) {
-	if unsafe.Pointer(pod.Spec()) != unsafe.Pointer(podSpec) {
-		panic("didn't cache pod from pre-filter")
-	}
-	if val, ok := state.Read(prefilterStateKey); !ok {
-		panic("didn't propagate state from filter")
-	} else if _, ok := val.(prefilterStateVal)["a"]; !ok {
-		panic("value lost propagating from filter")
-	}
-	return
 }