#### What type of PR is this?

/kind feature

#### What this PR does / why we need it:

This implement PreScore and changes NodeNumberPlugin to use it.

#### Which issue(s) this PR fixes:

N/A

#### Special notes for your reviewer:

Almost done: I need to do a review pass tomorrow.

#### Does this PR introduce a user-facing change?

NONE

#### What are the benchmark results of this change?

Performance is slightly slower because we always call prefilter. So, when we move logic to prescore, there's an additional host call roundtrip.

```benchstat
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e
                          │ before.txt  │              now.txt              │
                          │   sec/op    │   sec/op     vs base              │
Example_NodeNumber/New-12   63.43m ± 4%   64.32m ± 1%       ~ (p=0.394 n=6)
Example_NodeNumber/Run-12   51.34µ ± 2%   51.88µ ± 2%       ~ (p=0.058 n=6)
geomean                     1.805m        1.827m       +1.23%
```