@codefromthecrypt 2023-07-21
ok once this is in, I'll implement configuration to make our NodeNumberPlugin more realistic.