@codefromthecrypt 2023-07-22
What's needed after this is a config ABI similar to http-wasm. I'll add links on that for someone to follow-up on:

https://http-wasm.io/http-handler-abi/#get_config
https://github.com/http-wasm/http-wasm-guest-tinygo/blob/27f49b76b63759ad5694e32545dba595cd5fe9db/handler/host.go#L22-L25
https://github.com/http-wasm/http-wasm-guest-tinygo/blob/27f49b76b63759ad5694e32545dba595cd5fe9db/handler/internal/imports/imports.go#L10-L11
https://github.com/http-wasm/http-wasm-host-go/blob/990427a52ee421643c21142d462304c91415932b/handler/middleware.go#L251-L258
https://github.com/http-wasm/http-wasm-host-go/blob/990427a52ee421643c21142d462304c91415932b/handler/options.go#L25-L30

This would be used to complete the NodeNumberPlugin and would work if the config can be read inside the wasm guest, assigned by the host.